Move PyObject serialization to its own file
